add validation for mutation lock/unlock only for FPP

+++ backend/hct_mis_api/apps/payment/services/payment_plan_services.py
@@ -117,6 +117,9 @@ class PaymentPlanService:
         return self.payment_plan
 
     def lock(self) -> PaymentPlan:
+        if self.payment_plan.is_follow_up:
+            raise GraphQLError("Lock is available only for Follow-up Payment Plan")
+
         if not self.payment_plan.can_be_locked:
             raise GraphQLError("At least one valid Payment should exist in order to Lock the Payment Plan")
 
@@ -130,6 +133,9 @@ class PaymentPlanService:
         return self.payment_plan
 
     def unlock(self) -> PaymentPlan:
+        if self.payment_plan.is_follow_up:
+            raise GraphQLError("Unlock is available only for Follow-up Payment Plan")
+
         self.payment_plan.delivery_mechanisms.all().delete()
         self.payment_plan.status_unlock()
         self.payment_plan.update_population_count_fields()
